adsys_mount: handler_user_mounts: use Result methods instead of match
Most idiomatic Rust code usually tends to use those methods rather than
plain matches when it makes sense, as it makes composition a bit easier.
Here, the use of map_err() makes it clear that we're doing error
handling and nothing else.

Note that we're using the ? shortcut operator here to do an early return
after the error handling, so the final value is already unwrapped.