adsys_mount: handle_user_mounts: use Iterator::any for the final result
Using an external mutable boolean to track this can be error-prone, and
there's plethora of iterator methods in Rust to handle this kind of
things. This is preferred as you don't have to track down where the
variable is defined and then used.

The downside is that we go through the list twice, once for the logging
and another for the "already mounted" thing, but in terms of performance
that shouldn't really be impactful.