adsys_mount: handle_user_mounts: use variable shadowing
Its idiomatic in Rust, and the variable names are thus a bit more
meaningful than "mu". We know it's a mutex, it's its type ;).

We make use of a local scope to handle the clones necessary for the
closure without having to find a new name for those, as _clone names are
a bit tiresome.